Finally done, wow that was a lot of code

+++ code/modules/clothing/spacesuits/phorosian.dm
@@ -4,6 +4,7 @@
 	icon = 'icons/obj/clothing/species/phorosian/suits.dmi'
 	icon_state = "phorosiansuit"
 	item_state = "phorosiansuit"
+	item_flags = STOPPRESSUREDAMAGE | THICKMATERIAL | PHORONGUARD
 	allowed = list(/obj/item/weapon/tank)
 	desc = "A special containment suit designed to protect a phorosians volatile body from outside exposure."
 	armor = list(melee = 0, bullet = 0, laser = 0, energy = 0, bomb = 0, bio = 100, rad = 20)
@@ -22,10 +23,9 @@
 	icon = 'icons/obj/clothing/species/phorosian/hats.dmi'
 	icon_state = "phorosian_helmet0"
 	item_state = "phorosian_helmet0"
-	heat_protection = HEAD
+	flags_inv = HIDEMASK|HIDEEARS|HIDEEYES|BLOCKHAIR
+	item_flags = STOPPRESSUREDAMAGE | THICKMATERIAL | AIRTIGHT | PHORONGUARD
 	armor = list(melee = 0, bullet = 0, laser = 0, energy = 0, bomb = 0, bio = 100, rad = 20)
-	max_heat_protection_temperature = SPACE_SUIT_MAX_HEAT_PROTECTION_TEMPERATURE
-	siemens_coefficient = 0.4
 	species_restricted = list(SPECIES_PHOROSIAN)
 	light_overlay = "helmet_light"
 	sprite_sheets = list(
@@ -251,13 +251,13 @@
 	name = "Phorosian security suit"
 	icon_state = "phorosianSecurity_suit"
 	item_state = "phorosianSecurity_suit"
-	armor = list(melee = 60, bullet = 10, laser = 30, energy = 5, bomb = 45, bio = 100, rad = 10)
+	armor = list(melee = 40, bullet = 10, laser = 30, energy = 5, bomb = 45, bio = 100, rad = 10)
 
 /obj/item/clothing/head/helmet/space/phorosian/security
 	name = "Phorosian security helmet"
 	icon_state = "phorosianSecurity_helmet0"
 	item_state = "phorosianSecurity_helmet0"
-	armor = list(melee = 60, bullet = 10, laser = 30, energy = 5, bomb = 45, bio = 100, rad = 10)
+	armor = list(melee = 40, bullet = 10, laser = 30, energy = 5, bomb = 45, bio = 100, rad = 10)
 
 /obj/item/clothing/suit/space/phorosian/security/hos
 	name = "Phorosian head of security suit"
@@ -269,15 +269,17 @@
 	icon_state = "phorosianHoS_helmet0"
 	item_state = "phorosianHoS_helmet0"
 
-/obj/item/clothing/suit/space/phorosian/security/hop
+/obj/item/clothing/suit/space/phorosian/hop
 	name = "Phorosian head of personnel suit"
 	icon_state = "phorosianHoP"
 	item_state = "phorosianHoP"
+	armor = list(melee = 30, bullet = 15, laser = 40, energy = 10, bomb = 25, bio = 0, rad = 0)
 
-/obj/item/clothing/head/helmet/space/phorosian/security/hop
+/obj/item/clothing/head/helmet/space/phorosian/hop
 	name = "Phorosian head of personnel helmet"
 	icon_state = "phorosianHoP_helmet0"
 	item_state = "phorosianHoP_helmet0"
+	armor = list(melee = 30, bullet = 15, laser = 40, energy = 10, bomb = 25, bio = 0, rad = 0)
 
 /obj/item/clothing/suit/space/phorosian/security/captain
 	name = "Phorosian captain suit"
@@ -304,14 +306,18 @@
 	
 	
 /obj/item/device/phorosiansuit_changer //Can be used to change the type of plasmaman suit.
+	var/used = 0
 	name = "Phorosian suit adapter kit"
 	desc = "A device used to recolor and adapt a Phorosian containment suit to be more suited for the job they are assigned to."
-	icon_state = "modkit"
+	icon='icons/obj/storage.dmi'
+	icon_state = "inf_deployer"
 	w_class = 2
 	force = 0
 	throwforce = 0
 
 
+#define USED_ADAPT_HELM 1
+#define USED_ADAPT_SUIT 2
 
 /obj/item/device/phorosiansuit_changer/afterattack(atom/target, mob/user, proximity)
 	if(!proximity || !ishuman(user) || user.lying)
@@ -319,76 +325,87 @@
 	var/mob/living/carbon/human/H = user
 	var/suit=/obj/item/clothing/suit/space/phorosian
 	var/helm=/obj/item/clothing/head/helmet/space/phorosian
-	switch(H.mind.assigned_job.uid)
-		if("scientist","roboticist")
+	to_chat(H, "<span class='notice'>[H.mind.assigned_role]</span>")
+	switch(H.mind.assigned_role)
+		if("Scientist","Roboticist", "Xenobiologist")
 			suit=/obj/item/clothing/suit/space/phorosian/science
 			helm=/obj/item/clothing/head/helmet/space/phorosian/science
-		if("rd")
+		if("Research Director")
 			suit=/obj/item/clothing/suit/space/phorosian/science/rd
 			helm=/obj/item/clothing/head/helmet/space/phorosian/science/rd
-		if("engineer","mechanic")
+		if("Engineer","mechanic")
 			suit=/obj/item/clothing/suit/space/phorosian/engineer/
 			helm=/obj/item/clothing/head/helmet/space/phorosian/engineer/
-		if("chief")
+		if("Chief Engineer")
 			suit=/obj/item/clothing/suit/space/phorosian/engineer/ce
 			helm=/obj/item/clothing/head/helmet/space/phorosian/engineer/ce
-		if("atmostech")
+		if("Atmospheric Technician")
 			suit=/obj/item/clothing/suit/space/phorosian/atmostech
 			helm=/obj/item/clothing/head/helmet/space/phorosian/atmostech
-		if("pilot","judge","brigdoc","warden","detective","officer")
+		if("pilot","judge","brigdoc","Detective","Security Officer")
 			suit=/obj/item/clothing/suit/space/phorosian/security/
 			helm=/obj/item/clothing/head/helmet/space/phorosian/security/
-		if("hos")
+		if("Head of Security", "Warden")
 			suit=/obj/item/clothing/suit/space/phorosian/security/hos
 			helm=/obj/item/clothing/head/helmet/space/phorosian/security/hos
-		if("captain","nano","blueshield")
+		if("Captain","nano","blueshield")
 			suit=/obj/item/clothing/suit/space/phorosian/security/captain
 			helm=/obj/item/clothing/head/helmet/space/phorosian/security/captain
-		if("hop")
-			suit=/obj/item/clothing/suit/space/phorosian/security/hop
-			helm=/obj/item/clothing/head/helmet/space/phorosian/security/hop
-		if("doctor")
+		if("Head of Personnel")
+			suit=/obj/item/clothing/suit/space/phorosian/hop
+			helm=/obj/item/clothing/head/helmet/space/phorosian/hop
+		if("Medical Doctor", "Geneticist", "Psychiatrist")
 			suit=/obj/item/clothing/suit/space/phorosian/medical
 			helm=/obj/item/clothing/head/helmet/space/phorosian/medical
-		if("paramedic")
+		if("Paramedic")
 			suit=/obj/item/clothing/suit/space/phorosian/medical/paramedic
 			helm=/obj/item/clothing/head/helmet/space/phorosian/medical/paramedic
-		if("chemist")
+		if("Chemist")
 			suit=/obj/item/clothing/suit/space/phorosian/medical/chemist
 			helm=/obj/item/clothing/head/helmet/space/phorosian/medical/chemist
-		if("cmo")
+		if("Chief Medical Officer")
 			suit=/obj/item/clothing/suit/space/phorosian/medical/cmo
 			helm=/obj/item/clothing/head/helmet/space/phorosian/medical/cmo
-		if("chef","bartender")
+		if("Chef","Bartender")
 			suit=/obj/item/clothing/suit/space/phorosian/service
 			helm=/obj/item/clothing/head/helmet/space/phorosian/service
-		if("quartermaster","cargotech")
+		if("Quartermaster","Cargo Technician")
 			suit=/obj/item/clothing/suit/space/phorosian/cargo
 			helm=/obj/item/clothing/head/helmet/space/phorosian/cargo
-		if("miner")
+		if("Shaft Miner")
 			suit=/obj/item/clothing/suit/space/phorosian/miner
 			helm=/obj/item/clothing/head/helmet/space/phorosian/miner
-		if("botanist")
+		if("Gardener")
 			suit=/obj/item/clothing/suit/space/phorosian/botanist
 			helm=/obj/item/clothing/head/helmet/space/phorosian/botanist
-		if("chaplain")
+		if("Chaplain")
 			suit=/obj/item/clothing/suit/space/phorosian/chaplain
 			helm=/obj/item/clothing/head/helmet/space/phorosian/chaplain
-		if("janitor")
+		if("Janitor")
 			suit=/obj/item/clothing/suit/space/phorosian/janitor
 			helm=/obj/item/clothing/head/helmet/space/phorosian/janitor
-		if("civilian","lawyer")
+		if("Civilian","Internal Affairs Agent","Assistant","Librarian")
 			suit=/obj/item/clothing/suit/space/phorosian/assistant
 			helm=/obj/item/clothing/head/helmet/space/phorosian/assistant
 	
 	if(istype(target, /obj/item/clothing/head/helmet/space/phorosian))
+		if(used & USED_ADAPT_HELM)
+			to_chat(H, "<span class='notice'>The kit's helmet modifier has already been used.</span>")
+			return
 		H.equip_to_slot(new helm(H), slot_head)		
 		qdel(target)
 		to_chat(H, "<span class='notice'>You use the kit on [target], adapting it to suit your current job.</span>")
+		used |= USED_ADAPT_HELM
 	if (istype(target, /obj/item/clothing/suit/space/phorosian))
+		if(used & USED_ADAPT_SUIT)
+			to_chat(user, "<span class='notice'>The kit's suit modifier has already been used.</span>")
+			return
 		H.equip_to_slot(new suit(H), slot_wear_suit)
 		qdel(target)
 		to_chat(H, "<span class='notice'>You use the kit on [target], adapting it to suit your current job.</span>")
+		used |= USED_ADAPT_SUIT
 	return
 	to_chat(user, "<span class='warning'>You can't modify [target]!</span>")
-
+	
+#undef USED_ADAPT_HELM
+#undef USED_ADAPT_SUIT
